Remove OG doughnut v0 impl mv parity.rs -> mod.rs
