+++ Cargo.toml
@@ -8,7 +8,6 @@ license = "UNLICENSED"
 repository = "https://github.com/cennznet/doughnut-rs"
 
 [dependencies]
-bit_reverse = { version = "0.1.8", default-features = false }
 codec = { package = "parity-scale-codec", version = "^1.3.0", default-features = false, features = ["derive"] }
 primitive-types = { version = "<1.0", default-features = false }
 schnorrkel = { version = "^0.8.5", optional = true }
@@ -26,4 +25,3 @@ std = [
   "primitive-types/std",
   "schnorrkel",
 ]
-
